requirements.txt
# Database
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
mysqlclient>=2.2.0

# Machine Learning
scikit-learn>=1.3.0
joblib>=1.3.0

# Data Processing & Analysis
numpy>=1.24.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0 